+++ app/client/cypress/e2e/Regression/ClientSide/Git/GitAutocommit_spec.ts
@@ -21,9 +21,6 @@ describe(
   },
   function () {
     it("Check if autocommit progress bar is visible and network requests are properly called", function () {
-      featureFlagIntercept({
-        release_git_autocommit_feature_enabled: true,
-      });
       agHelper.GenerateUUID();
       cy.get("@guid").then((uid) => {
         wsName = "GitAC-" + uid;
