+++ app/client/src/ce/entities/FeatureFlag.ts
@@ -21,8 +21,6 @@ export const FEATURE_FLAG = {
     "license_git_branch_protection_enabled",
   license_git_continuous_delivery_enabled:
     "license_git_continuous_delivery_enabled",
-  release_git_autocommit_feature_enabled:
-    "release_git_autocommit_feature_enabled",
   license_widget_rtl_support_enabled: "license_widget_rtl_support_enabled",
   ab_one_click_learning_popover_enabled:
     "ab_one_click_learning_popover_enabled",
@@ -68,7 +66,6 @@ export const DEFAULT_FEATURE_FLAG_VALUE: FeatureFlags = {
   release_drag_drop_building_blocks_enabled: false,
   release_table_cell_label_value_enabled: false,
   license_git_branch_protection_enabled: false,
-  release_git_autocommit_feature_enabled: false,
   license_git_continuous_delivery_enabled: false,
   license_widget_rtl_support_enabled: false,
   ab_one_click_learning_popover_enabled: false,
