+++ app/server/appsmith-git/src/main/java/com/appsmith/git/files/operations/FileOperationsCEv2Impl.java
@@ -1,15 +1,11 @@
 package com.appsmith.git.files.operations;
 
-import com.appsmith.external.annotations.FeatureFlagged;
-import com.appsmith.external.enums.FeatureFlagEnum;
-import com.appsmith.external.git.GitExecutor;
 import com.appsmith.external.git.constants.GitSpan;
 import com.appsmith.external.git.operations.FileOperationsCE;
 import com.appsmith.external.helpers.ObservationHelper;
 import com.appsmith.external.models.ApplicationGitReference;
 import com.appsmith.external.models.BaseDomain;
 import com.appsmith.external.views.Git;
-import com.appsmith.git.configurations.GitServiceConfig;
 import com.appsmith.git.constants.CommonConstants;
 import com.appsmith.util.SerializationUtils;
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -20,25 +16,32 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ObjectReader;
 import com.fasterxml.jackson.databind.ObjectWriter;
 import com.fasterxml.jackson.databind.node.ObjectNode;
-import com.google.gson.GsonBuilder;
 import io.micrometer.tracing.Span;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.io.FileUtils;
 import org.json.JSONObject;
-import org.springframework.context.annotation.Import;
 import org.springframework.stereotype.Component;
+import reactor.core.publisher.Mono;
 
 import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
+import java.nio.file.DirectoryNotEmptyException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.nio.file.attribute.BasicFileAttributes;
+import java.nio.file.attribute.FileTime;
+import java.time.Instant;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
+import java.util.Set;
+import java.util.stream.Stream;
 
 import static com.appsmith.external.git.constants.ce.GitConstantsCE.GitMetricConstantsCE.METADATA;
 import static com.appsmith.external.git.constants.ce.GitConstantsCE.GitMetricConstantsCE.RESOURCE_TYPE;
@@ -47,8 +50,7 @@ import static com.appsmith.external.git.constants.ce.GitConstantsCE.GitMetricCon
 @Slf4j
 @Getter
 @Component
-@Import({GitServiceConfig.class})
-public class FileOperationsCEv2Impl extends FileOperationsCEImpl implements FileOperationsCE {
+public class FileOperationsCEv2Impl implements FileOperationsCE {
 
     protected final ObjectMapper objectMapper;
     protected final ObjectReader objectReader;
@@ -56,13 +58,7 @@ public class FileOperationsCEv2Impl extends FileOperationsCEImpl implements File
 
     private final ObservationHelper observationHelper;
 
-    public FileOperationsCEv2Impl(
-            GitServiceConfig gitServiceConfig,
-            GitExecutor gitExecutor,
-            GsonBuilder gsonBuilder,
-            PrettyPrinter prettyPrinter,
-            ObservationHelper observationHelper) {
-        super(gitServiceConfig, gitExecutor, gsonBuilder, observationHelper);
+    public FileOperationsCEv2Impl(PrettyPrinter prettyPrinter, ObservationHelper observationHelper) {
 
         this.objectMapper = SerializationUtils.getBasicObjectMapper(prettyPrinter);
 
@@ -73,7 +69,6 @@ public class FileOperationsCEv2Impl extends FileOperationsCEImpl implements File
         this.observationHelper = observationHelper;
     }
 
-    @FeatureFlagged(featureFlagName = FeatureFlagEnum.release_git_autocommit_feature_enabled)
     @Override
     public void saveMetadataResource(ApplicationGitReference applicationGitReference, Path baseRepo) {
         ObjectNode metadata = objectMapper.valueToTree(applicationGitReference.getMetadata());
@@ -81,7 +76,6 @@ public class FileOperationsCEv2Impl extends FileOperationsCEImpl implements File
         saveResource(metadata, baseRepo.resolve(CommonConstants.METADATA + CommonConstants.JSON_EXTENSION));
     }
 
-    @FeatureFlagged(featureFlagName = FeatureFlagEnum.release_git_autocommit_feature_enabled)
     @Override
     public void saveWidgets(JSONObject sourceEntity, String resourceName, Path path) {
         Span span = observationHelper.createSpan(GitSpan.FILE_WRITE);
@@ -101,7 +95,6 @@ public class FileOperationsCEv2Impl extends FileOperationsCEImpl implements File
         }
     }
 
-    @FeatureFlagged(featureFlagName = FeatureFlagEnum.release_git_autocommit_feature_enabled)
     @Override
     public boolean writeToFile(Object sourceEntity, Path path) throws IOException {
         Span span = observationHelper.createSpan(GitSpan.FILE_WRITE);
@@ -126,7 +119,6 @@ public class FileOperationsCEv2Impl extends FileOperationsCEImpl implements File
      * @param filePath file on which the read operation will be performed
      * @return resource stored in the JSON file
      */
-    @FeatureFlagged(featureFlagName = FeatureFlagEnum.release_git_autocommit_feature_enabled)
     @Override
     public Object readFile(Path filePath) {
         Span span = observationHelper.createSpan(GitSpan.FILE_READ);
@@ -150,7 +142,6 @@ public class FileOperationsCEv2Impl extends FileOperationsCEImpl implements File
      * @param directoryPath directory path for files on which read operation will be performed
      * @return resources stored in the directory
      */
-    @FeatureFlagged(featureFlagName = FeatureFlagEnum.release_git_autocommit_feature_enabled)
     @Override
     public Map<String, Object> readFiles(Path directoryPath, String keySuffix) {
         Map<String, Object> resource = new HashMap<>();
@@ -171,7 +162,6 @@ public class FileOperationsCEv2Impl extends FileOperationsCEImpl implements File
         return resource;
     }
 
-    @FeatureFlagged(featureFlagName = FeatureFlagEnum.release_git_autocommit_feature_enabled)
     @Override
     public Integer getFileFormatVersion(Object metadata) {
         if (metadata == null) {
@@ -182,7 +172,6 @@ public class FileOperationsCEv2Impl extends FileOperationsCEImpl implements File
         return fileFormatVersion;
     }
 
-    @FeatureFlagged(featureFlagName = FeatureFlagEnum.release_git_autocommit_feature_enabled)
     @Override
     public JSONObject getMainContainer(Object pageJson) {
         JsonNode pageJSON = objectMapper.valueToTree(pageJson);
@@ -191,6 +180,147 @@ public class FileOperationsCEv2Impl extends FileOperationsCEImpl implements File
                     pageJSON.get("unpublishedPage").get("layouts").get(0).get("dsl")));
         } catch (JsonProcessingException e) {
             throw new RuntimeException(e);
+        }
+    }
+
+    /**
+     * This method will be used to store the DB resource to JSON file
+     *
+     * @param sourceEntity resource extracted from DB to be stored in file
+     * @param path         file path where the resource to be stored
+     * @return if the file operation is successful
+     */
+    @Override
+    public boolean saveResource(Object sourceEntity, Path path) {
+        try {
+            Files.createDirectories(path.getParent());
+            return writeToFile(sourceEntity, path);
+        } catch (IOException e) {
+            log.error("Error while writing resource to file {} with {}", path, e.getMessage());
+            log.debug(e.getMessage());
+        }
+        return false;
+    }
+
+    /**
+     * This method will delete the JSON resource available in local git directory on subsequent commit made after the
+     * deletion of respective resource from DB
+     *
+     * @param validResources    resources those are still available in DB
+     * @param resourceDirectory directory which needs to be scanned for possible file deletion operations
+     */
+    @Override
+    public void scanAndDeleteFileForDeletedResources(Set<String> validResources, Path resourceDirectory) {
+        // Scan resource directory and delete any unwanted file if present
+        // unwanted file : corresponding resource from DB has been deleted
+        if (resourceDirectory.toFile().exists()) {
+            try (Stream<Path> paths = Files.walk(resourceDirectory)) {
+                paths.filter(pathLocal -> Files.isRegularFile(pathLocal)
+                                && !validResources.contains(
+                                        pathLocal.getFileName().toString()))
+                        .forEach(this::deleteFile);
+            } catch (IOException e) {
+                log.error("Error while scanning directory: {}, with error {}", resourceDirectory, e.getMessage());
+            }
+        }
+    }
+
+    /**
+     * This method will delete the JSON resource directory available in local git directory on subsequent commit made after the
+     * deletion of respective resource from DB
+     *
+     * @param validResources    resources those are still available in DB
+     * @param resourceDirectory directory which needs to be scanned for possible file deletion operations
+     */
+    @Override
+    public void scanAndDeleteDirectoryForDeletedResources(Set<String> validResources, Path resourceDirectory) {
+        // Scan resource directory and delete any unwanted directory if present
+        // unwanted directory : corresponding resource from DB has been deleted
+        if (resourceDirectory.toFile().exists()) {
+            try (Stream<Path> paths = Files.walk(resourceDirectory, 1)) {
+                paths.filter(path -> Files.isDirectory(path)
+                                && !path.equals(resourceDirectory)
+                                && !validResources.contains(path.getFileName().toString()))
+                        .forEach(this::deleteDirectory);
+            } catch (IOException e) {
+                log.error("Error while scanning directory {} with error {}", resourceDirectory, e.getMessage());
+            }
+        }
+    }
+
+    /**
+     * This method will delete the directory and all its contents
+     *
+     * @param directory
+     */
+    @Override
+    public void deleteDirectory(Path directory) {
+        if (directory.toFile().exists()) {
+            try {
+                FileUtils.deleteDirectory(directory.toFile());
+            } catch (IOException e) {
+                log.error("Unable to delete directory for path {} with message {}", directory, e.getMessage());
+            }
+        }
+    }
+
+    /**
+     * This method will delete the file from local repo
+     *
+     * @param filePath file that needs to be deleted
+     */
+    @Override
+    public void deleteFile(Path filePath) {
+        try {
+            Files.deleteIfExists(filePath);
+        } catch (DirectoryNotEmptyException e) {
+            log.error("Unable to delete non-empty directory at {} with cause", filePath, e.getMessage());
+        } catch (IOException e) {
+            log.error("Unable to delete file {} with {}", filePath, e.getMessage());
+        }
+    }
+
+    /**
+     * This method will read the content of the file as a plain text and does not apply the gson to json transformation
+     *
+     * @param filePath file path for files on which read operation will be performed
+     * @return content of the file in the path
+     */
+    @Override
+    public String readFileAsString(Path filePath) {
+        Span span = observationHelper.createSpan(GitSpan.FILE_READ);
+        observationHelper.startSpan(span, true);
+        String data = CommonConstants.EMPTY_STRING;
+        try {
+            data = FileUtils.readFileToString(filePath.toFile(), "UTF-8");
+        } catch (IOException e) {
+            log.error("Error while reading the file from git repo {} ", e.getMessage());
+        } finally {
+            observationHelper.endSpan(span, true);
+        }
+        return data;
+    }
+
+    @Override
+    public Mono<Long> deleteIndexLockFile(Path path, int validTimeInSeconds) {
+        // Check the time created of the index.lock file
+        // If the File is stale for more than validTime, then delete the file
+        try {
+            BasicFileAttributes attr = Files.readAttributes(path, BasicFileAttributes.class);
+            FileTime fileTime = attr.creationTime();
+            Instant now = Instant.now();
+            Instant validCreateTime = now.minusSeconds(validTimeInSeconds);
+            if (fileTime.toInstant().isBefore(validCreateTime)) {
+                // Add base repo path
+                path = Paths.get(path + ".lock");
+                deleteFile(path);
+                return Mono.just(now.minusMillis(fileTime.toMillis()).getEpochSecond());
+            } else {
+                return Mono.just(0L);
+            }
+        } catch (IOException ex) {
+            log.error("Error reading index.lock file: {}", ex.getMessage());
+            return Mono.just(0L);
         }
     }
 }
