+++ app/server/appsmith-git/src/main/java/com/appsmith/git/files/operations/FileOperationsImpl.java
@@ -1,25 +1,15 @@
 package com.appsmith.git.files.operations;
 
-import com.appsmith.external.git.GitExecutor;
 import com.appsmith.external.git.operations.FileOperations;
 import com.appsmith.external.helpers.ObservationHelper;
-import com.appsmith.git.configurations.GitServiceConfig;
 import com.fasterxml.jackson.core.PrettyPrinter;
-import com.google.gson.GsonBuilder;
-import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Component;
 
-@Component
 @Primary
-@Import({GitServiceConfig.class})
+@Component
 public class FileOperationsImpl extends FileOperationsCEv2Impl implements FileOperations {
-    public FileOperationsImpl(
-            GitServiceConfig gitServiceConfig,
-            GitExecutor gitExecutor,
-            GsonBuilder gsonBuilder,
-            PrettyPrinter prettyPrinter,
-            ObservationHelper observationHelper) {
-        super(gitServiceConfig, gitExecutor, gsonBuilder, prettyPrinter, observationHelper);
+    public FileOperationsImpl(PrettyPrinter prettyPrinter, ObservationHelper observationHelper) {
+        super(prettyPrinter, observationHelper);
     }
 }
