+++ app/server/appsmith-interfaces/src/main/java/com/appsmith/external/enums/FeatureFlagEnum.java
@@ -14,16 +14,6 @@ public enum FeatureFlagEnum {
     APP_NAVIGATION_LOGO_UPLOAD,
     release_embed_hide_share_settings_enabled,
     rollout_datasource_test_rate_limit_enabled,
-    release_git_autocommit_feature_enabled,
-    /**
-     * Since checking eligibility for autocommit is an expensive operation,
-     * We want to roll out this feature on cloud in a controlled manner.
-     * We could have used the autocommit flag itself, however it is on tenant level,
-     * and it can't be moved to user level due to its usage on non-reactive code paths.
-     * We would keep the main autocommit flag false on production for the version <= testing versions,
-     * and turn it to true for later versions
-     * We would remove this feature flag once the testing is complete.
-     */
-    release_git_autocommit_eligibility_enabled,
+
     // Add EE flags below this line, to avoid conflicts.
 }
