+++ app/server/appsmith-server/src/main/java/com/appsmith/server/git/autocommit/helpers/AutoCommitEligibilityHelperImpl.java
@@ -25,8 +25,7 @@ import static java.lang.Boolean.TRUE;
 @Primary
 @Component
 @RequiredArgsConstructor
-public class AutoCommitEligibilityHelperImpl extends AutoCommitEligibilityHelperFallbackImpl
-        implements AutoCommitEligibilityHelper {
+public class AutoCommitEligibilityHelperImpl implements AutoCommitEligibilityHelper {
 
     private final CommonGitFileUtils commonGitFileUtils;
     private final DSLMigrationUtils dslMigrationUtils;
