+++ app/server/appsmith-server/src/test/java/com/appsmith/server/git/autocommit/helpers/GitAutoCommitHelperImplTest.java
@@ -1,6 +1,5 @@
 package com.appsmith.server.git.autocommit.helpers;
 
-import com.appsmith.external.enums.FeatureFlagEnum;
 import com.appsmith.server.acl.AclPermission;
 import com.appsmith.server.applications.base.ApplicationService;
 import com.appsmith.server.domains.Application;
@@ -14,7 +13,6 @@ import com.appsmith.server.git.autocommit.AutoCommitEventHandler;
 import com.appsmith.server.git.common.CommonGitService;
 import com.appsmith.server.helpers.GitPrivateRepoHelper;
 import com.appsmith.server.helpers.RedisUtils;
-import com.appsmith.server.services.FeatureFlagService;
 import com.appsmith.server.services.UserDataService;
 import com.appsmith.server.solutions.ApplicationPermission;
 import lombok.extern.slf4j.Slf4j;
@@ -42,9 +40,6 @@ import static org.mockito.ArgumentMatchers.eq;
 @Slf4j
 @DirtiesContext
 public class GitAutoCommitHelperImplTest {
-
-    @MockBean
-    FeatureFlagService featureFlagService;
 
     @MockBean
     AutoCommitEventHandler autoCommitEventHandler;
@@ -81,35 +76,11 @@ public class GitAutoCommitHelperImplTest {
     }
 
     @Test
-    public void autoCommitApplication_WhenFeatureFlagIsDisabled_AutoCommitNotTriggered() {
-        Application application = new Application();
-        application.setGitApplicationMetadata(new GitArtifactMetadata());
-
-        Mockito.when(featureFlagService.check(FeatureFlagEnum.release_git_autocommit_feature_enabled))
-                .thenReturn(Mono.just(Boolean.FALSE));
-        Mockito.when(applicationService.findById(anyString(), any(AclPermission.class)))
-                .thenReturn(Mono.just(application));
-        Mockito.when(gitPrivateRepoHelper.isBranchProtected(any(GitArtifactMetadata.class), anyString()))
-                .thenReturn(Mono.just(Boolean.FALSE));
-        Mockito.when(applicationService.findByBranchNameAndBaseApplicationId(
-                        anyString(), anyString(), any(AclPermission.class)))
-                .thenReturn(Mono.just(application));
-
-        StepVerifier.create(gitAutoCommitHelper.autoCommitClientMigration(defaultApplicationId, branchName))
-                .assertNext(aBoolean -> {
-                    assertThat(aBoolean).isFalse();
-                })
-                .verifyComplete();
-    }
-
-    @Test
     public void autoCommitApplication_WhenBranchIsProtected_AutoCommitNotTriggered() {
         Application application = new Application();
         application.setId(defaultApplicationId);
         application.setGitApplicationMetadata(new GitArtifactMetadata());
 
-        Mockito.when(featureFlagService.check(FeatureFlagEnum.release_git_autocommit_feature_enabled))
-                .thenReturn(Mono.just(Boolean.TRUE));
         Mockito.when(applicationService.findById(defaultApplicationId, applicationPermission.getEditPermission()))
                 .thenReturn(Mono.just(application));
         Mockito.when(gitPrivateRepoHelper.isBranchProtected(any(GitArtifactMetadata.class), eq(branchName)))
@@ -132,11 +103,8 @@ public class GitAutoCommitHelperImplTest {
         GitArtifactMetadata metadata = new GitArtifactMetadata();
         metadata.setAutoCommitConfig(new AutoCommitConfig());
         metadata.getAutoCommitConfig().setEnabled(Boolean.FALSE);
-
         application.setGitApplicationMetadata(metadata);
 
-        Mockito.when(featureFlagService.check(FeatureFlagEnum.release_git_autocommit_feature_enabled))
-                .thenReturn(Mono.just(Boolean.TRUE));
         Mockito.when(applicationService.findById(defaultApplicationId, applicationPermission.getEditPermission()))
                 .thenReturn(Mono.just(application));
         Mockito.when(applicationService.findByBranchNameAndBaseApplicationId(
@@ -162,8 +130,6 @@ public class GitAutoCommitHelperImplTest {
                 .thenReturn(Mono.just(application));
         Mockito.when(gitPrivateRepoHelper.isBranchProtected(any(GitArtifactMetadata.class), eq(branchName)))
                 .thenReturn(Mono.just(Boolean.FALSE));
-        Mockito.when(featureFlagService.check(FeatureFlagEnum.release_git_autocommit_feature_enabled))
-                .thenReturn(Mono.just(Boolean.TRUE));
         Mockito.when(applicationService.findByBranchNameAndBaseApplicationId(
                         anyString(), anyString(), any(AclPermission.class)))
                 .thenReturn(Mono.just(application));
@@ -190,11 +156,8 @@ public class GitAutoCommitHelperImplTest {
         gitAuth.setPrivateKey("private-key");
         gitAuth.setPublicKey("public-key");
         metaData.setGitAuth(gitAuth);
-
         application.setGitApplicationMetadata(metaData);
 
-        Mockito.when(featureFlagService.check(FeatureFlagEnum.release_git_autocommit_feature_enabled))
-                .thenReturn(Mono.just(Boolean.TRUE));
         Mockito.when(applicationService.findById(defaultApplicationId, applicationPermission.getEditPermission()))
                 .thenReturn(Mono.just(application));
         Mockito.when(applicationService.findByBranchNameAndBaseApplicationId(
@@ -291,11 +254,8 @@ public class GitAutoCommitHelperImplTest {
         GitArtifactMetadata metadata = new GitArtifactMetadata();
         metadata.setAutoCommitConfig(new AutoCommitConfig());
         metadata.getAutoCommitConfig().setEnabled(Boolean.TRUE);
-
         application.setGitApplicationMetadata(metadata);
 
-        Mockito.when(featureFlagService.check(FeatureFlagEnum.release_git_autocommit_feature_enabled))
-                .thenReturn(Mono.just(Boolean.TRUE));
         Mockito.when(applicationService.findById(defaultApplicationId, applicationPermission.getEditPermission()))
                 .thenReturn(Mono.just(application));
         Mockito.when(applicationService.findByBranchNameAndBaseApplicationId(
@@ -316,31 +276,6 @@ public class GitAutoCommitHelperImplTest {
     }
 
     @Test
-    public void autoCommitApplication_WhenServerMigrationFlagIsFalse_returnsFalse() {
-        Application application = new Application();
-        application.setWorkspaceId("sample-workspace-id");
-        GitArtifactMetadata metaData = new GitArtifactMetadata();
-        metaData.setRepoName("test-repo-name");
-        metaData.setDefaultApplicationId(defaultApplicationId);
-        metaData.setBranchName(branchName);
-
-        GitAuth gitAuth = new GitAuth();
-        gitAuth.setPrivateKey("private-key");
-        gitAuth.setPublicKey("public-key");
-        metaData.setGitAuth(gitAuth);
-
-        application.setGitApplicationMetadata(metaData);
-
-        Mockito.when(featureFlagService.check(FeatureFlagEnum.release_git_autocommit_feature_enabled))
-                .thenReturn(Mono.just(Boolean.FALSE));
-
-        StepVerifier.create(gitAutoCommitHelper.autoCommitServerMigration(defaultApplicationId, branchName))
-                .assertNext(isAutoCommitPublished -> {
-                    assertThat(isAutoCommitPublished).isFalse();
-                });
-    }
-
-    @Test
     public void autoCommitApplication_WhenServerRequiresMigration_successIfEverythingIsRight() {
         Application application = new Application();
         application.setWorkspaceId("sample-workspace-id");
@@ -355,9 +290,6 @@ public class GitAutoCommitHelperImplTest {
         metaData.setGitAuth(gitAuth);
 
         application.setGitApplicationMetadata(metaData);
-
-        Mockito.when(featureFlagService.check(FeatureFlagEnum.release_git_autocommit_feature_enabled))
-                .thenReturn(Mono.just(Boolean.TRUE));
 
         Mockito.when(applicationService.findById(anyString(), any(AclPermission.class)))
                 .thenReturn(Mono.just(application));
@@ -401,11 +333,7 @@ public class GitAutoCommitHelperImplTest {
         gitAuth.setPrivateKey("private-key");
         gitAuth.setPublicKey("public-key");
         metaData.setGitAuth(gitAuth);
-
         application.setGitApplicationMetadata(metaData);
-
-        Mockito.when(featureFlagService.check(FeatureFlagEnum.release_git_autocommit_feature_enabled))
-                .thenReturn(Mono.just(Boolean.TRUE));
 
         Mockito.when(applicationService.findById(anyString(), any(AclPermission.class)))
                 .thenReturn(Mono.just(application));
